Separate stage definitions from runBenchmark.
